Modify CCD-bus idle timer.
Timer 3 prescaler changed to 1, bit-sensitivity improved by 1 bit, still not good enough.

Somehow 13 bits (+1 stop bit) are needed to sense message endings. Any smaller (should be 10) and the code considers every byte an independent message. This needs more work.

+++ src/CCDLibrary.h
@@ -93,7 +93,7 @@ class CCDLibrary
         uint8_t _busIdleBits;
         bool _verifyRxChecksum;
         bool _calculateTxChecksum;
-        uint8_t _calculatedOCRAValue;
+        uint16_t _calculatedOCRAValue;
         void serialInit(uint16_t ubrr);
         void busIdleTimerInit();
         void busIdleTimerStart();
